check for destroyPersistentStore. Extract utils to own file
